+++ angel_serialize_generator/test/models/book.g.dart
@@ -6,6 +6,7 @@ part of angel_serialize.test.models.book;
 // JsonModelGenerator
 // **************************************************************************
 
+@generatedSerializable
 class Book extends _Book {
   Book(
       {this.id,