+++ angel_serialize_generator/test/models/book.serializer.g.dart
@@ -19,12 +19,12 @@ abstract class BookSerializer {
         createdAt: map['created_at'] != null
             ? (map['created_at'] is DateTime
                 ? (map['created_at'] as DateTime)
-                : DateTime.parse(map['created_at']))
+                : DateTime.parse(map['created_at'].toString()))
             : null,
         updatedAt: map['updated_at'] != null
             ? (map['updated_at'] is DateTime
                 ? (map['updated_at'] as DateTime)
-                : DateTime.parse(map['updated_at']))
+                : DateTime.parse(map['updated_at'].toString()))
             : null);
   }